account management

+++ user.h
@@ -7,18 +7,18 @@ class User {
         int accountNumber;
         std::string name;
         int pin;
-        double balance;
+        size_t balance;
 
     public:
-        User(int accNum, std::string name,int pin,double balance);
+        User(int accNum, std::string name,int pin,size_t balance);
 
         void setName(std::string name);
         void setPin(int pin);
-        void deposit(double amount);
-        void withdraw(double amount);   
+        void deposit(size_t amount);
+        void withdraw(size_t amount);   
 
         int getAccountNumber() const;
         std::string getName() const;
         int getPin() const;
-        double getBalance() const;
+        size_t getBalance() const;
 };
